+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 *.local
+dist
 
 # vscode
 /.vscode/
