+++ src/views/home/index.js
@@ -151,6 +151,32 @@ let Home = {
 
       manipulateSlidesClasses(correctSlideIndex);
     });
+
+    //Consuming API Brands
+    let brands = document.querySelector("#marca");
+    const url = "https://e-carros-api.herokuapp.com/brands";
+    
+    function getBrands() {
+      fetch(url)
+        .then((res) => res.json())
+        .then((response) => {
+          brands.innerHTML = `
+            <ul class="brandList">
+              ${response.map(
+                (element, index) =>
+                  `
+                    <li class="brandItem" key="${index}">
+                      <a href="#" title="${element.name}">
+                        <img src="${element.logo}" width="80px" height="auto"/>
+                      </a>
+                    </li>
+                  `
+              ).join("")}
+            </ul>`;
+        });
+    }
+    getBrands();
+    
   },
 };
